% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_recipes.R
\name{add_recipes}
\alias{add_recipes}
\title{Add a list of recipes to a workflow}
\usage{
add_recipes(.workflow, .recipes, ...)
}
\arguments{
\item{.workflow}{A \code{workflow} object.}

\item{.recipes}{A list of \code{recipe} objects.}

\item{...}{Arguments passed on to \code{workflows::add_recipe()}.}
}
\value{
A \code{wflist} of replicate workflows, one for each provided recipe.
}
\description{
This function replicates the provided workflow, adding one of the
provided recipes to each replicate.
}
