% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.cv_thresholds_results}
\alias{autoplot.cv_thresholds_results}
\alias{autoplot.cv_thresholds}
\title{Plot cross-validated threshold metrics}
\usage{
\method{autoplot}{cv_thresholds_results}(object, ...)

\method{autoplot}{cv_thresholds}(object, ...)
}
\arguments{
\item{object}{The data frame produced by \code{cv_threshold_perf()} or
by \code{cv_threshold_perf() \%>\% collect_metrics()}.}

\item{...}{Not currently used.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot cross-validated threshold metrics
}
\section{Functions}{
\itemize{
\item \code{autoplot.cv_thresholds}: Can be used directly on the output
of \code{cv_threshold_perf()}.
}}

