% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_metrics.R
\name{collect_metrics.cv_thresholds}
\alias{collect_metrics.cv_thresholds}
\title{Summarize threshold metrics across folds}
\usage{
\method{collect_metrics}{cv_thresholds}(x, ...)
}
\arguments{
\item{x}{The data.frame produced by \code{cv_threshold_perf()}.}

\item{...}{Not currently used.}
}
\value{
A \code{data.frame} with class \code{cv_threshold_results}.
}
\description{
This function calculates means and standard deviations for each
metric at each threshold value across all folds.

See \code{?cv_threshold_perf} and \code{probably::threshold_perf} for details.

Can be piped into \code{autoplot()}.
}
