% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_predictors.R
\name{count_predictors}
\alias{count_predictors}
\title{Get the number of predictors}
\usage{
count_predictors(x)
}
\arguments{
\item{x}{A \code{recipe} object or a \code{workflow} object with an attached recipe.}
}
\value{
An integer.
}
\description{
Returns the number of variables with the "predictor" role in a recipe or workflow.
}
