% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validate.R
\name{cross_validate}
\alias{cross_validate}
\alias{cross_validate.workflow}
\alias{cross_validate.wflist}
\title{Cross-validate your workflows}
\usage{
cross_validate(x, resamples, outcome, mode)

\method{cross_validate}{workflow}(x, resamples, outcome, mode)

\method{cross_validate}{wflist}(x, resamples, outcome, mode)
}
\arguments{
\item{x}{A \code{workflow} object or a \code{wflist} of \code{workflow} objects created by \code{wf_list()}.}

\item{resamples}{A resample \code{rset} created from an \code{rsample} function such as \code{rsample::vfold_cv()}.}

\item{outcome}{A character string containing the column identifier for the outcome variable.}

\item{mode}{Either "regression" or "classification".}
}
\value{
A \code{data.frame} containing the resulting metrics and their standard errors. When
\code{x} is a \code{wflist}, there is a row for each workflow's metrics.
}
\description{
This is a wrapper around \code{tune::fit_resamples()} that returns correct standard errors
for the cross-validation metrics.
}
\details{
Setting mode = "regression" calculates the cross-validated mean square error.

Setting mode = "classification" calculates the cross-validated accuracy.

For some reason, \code{tune::fit_resamples() \%>\% tune::collect_metrics()} does not return
the correct standard errors for its metrics. It's a bummer since the \code{tune} functions
are so easy to use.

When called on a \code{wflist} of workflows, this function also provides information on how
the workflows compare to each other. In "regression" mode, the difference between each
workflow's mean square error and the best mean square error is shown. In
"classification" mode, the proportion of rows where each workflow disagrees with the
workflow with the highest accuracy is shown. Both of these additional metrics come
with standard errors.
}
\section{Methods (by class)}{
\itemize{
\item \code{workflow}: Cross-validates a single workflow.

\item \code{wflist}: Cross-validates multiple workflows and generates comparison metrics.
}}

