% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_threshold_perf.R
\name{cv_threshold_perf}
\alias{cv_threshold_perf}
\title{Title Generate cross-validated performance metrics across probability thresholds}
\usage{
cv_threshold_perf(object, resamples, outcome, thresholds = NULL, ...)
}
\arguments{
\item{object}{A \code{workflow} object.}

\item{resamples}{A resample \code{rset} created from an \code{rsample} function such
as \code{rsample::vfold_cv()}.}

\item{outcome}{A character string containing the column identifier for the
outcome variable.}

\item{thresholds}{A numeric vector of values for the probability threshold.
If unspecified, a series of values between 0.5 and 1.0 are used. Note: if
this argument is used, it must be named.}

\item{...}{Additional arguments passed on to \code{probably::threshold_perf()}.}
}
\value{
A \code{data.frame} containing the resulting metrics for each threshold
and fold.
}
\description{
This function calculates out-of-sample performance characteristics for class
probability thresholds for each provided split.
}
\details{
For each training-testing pair, this function fits a model on the training
set and generates class probability predictions for the testing set.
\code{probably::threshold_perf()} is called on those out-of-sample predictions.

See \code{?probably::threshold_perf} for details.
}
