% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_replicates.R
\name{default_replicates}
\alias{default_replicates}
\title{Estimate how many down-sampling replicates to use}
\usage{
default_replicates(x, outcome, coverage = 0.8)
}
\arguments{
\item{x}{A \code{data.frame} containing the data to be down-sampled.}

\item{outcome}{A variable specified as:
\itemize{
\item a literal variable name
\item a positive integer, giving the position counting from the left
\item a negative integer, giving the position counting from the right.
}}

\item{coverage}{The desired proportion of the over-represented class
to be represented in the aggregated down-samples.}
}
\value{
A positive integer
}
\description{
Given data with an under-represented binary outcome, this function
estimates the number of times the data would need to be down-sampled
for the union of the resulting down-samples to contain a given
proportion of the over-represented class.
}
