% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess.R
\name{guess}
\alias{guess}
\title{Predict unseen data}
\usage{
guess(object, new_data, mode)
}
\arguments{
\item{object}{A \code{workflow} object.}

\item{new_data}{A \code{data.frame} containing data with which to predict.}

\item{mode}{Either "regression" or "classification".}
}
\value{
A \code{data.frame} (new_data with new columns containing the predictions).
}
\description{
An enhanced \code{predict()} that returns the provided data, predictions for that data,
and additional information about the predictions.
}
\details{
Setting mode = "regression" provides prediction intervals and setting
mode = "classification" provides class probabilities.
}
