% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_resamples.R
\name{predict_resamples}
\alias{predict_resamples}
\title{Predict resampled data using multiple models}
\usage{
predict_resamples(objects, resamples)
}
\arguments{
\item{objects}{A named list of \code{workflow} objects.}

\item{resamples}{A resample \code{rset} created from an \code{rsample} function such as \code{rsample::vfold_cv()}.}
}
\value{
A \code{data.frame} with one column for each fit's predictions.
}
\description{
Fits a list of \code{workflow} objects to resampled data. Its output can be used to train
higher models in a stack.
}
