% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R, R/collect_metrics.R, R/fit.R,
%   R/predict.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{autoplot}
\alias{collect_metrics}
\alias{fit}
\alias{predict}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{fit}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}

  \item{stats}{\code{\link[stats]{predict}}}

  \item{tune}{\code{\link[tune:collect_predictions]{collect_metrics}}}
}}

