% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_importance.R
\name{variable_importance}
\alias{variable_importance}
\title{Get variable importance metrics}
\usage{
variable_importance(object)
}
\arguments{
\item{object}{A fitted \code{workflow} object.}
}
\value{
Return type depends on model engine.
}
\description{
This function calls the appropriate variable importance function
for the model engine used. See the Details section for the
of engines that are currently supported.
}
\details{
Currently only the \code{ranger} and \code{xgboost} engines are supported.
}
