% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflist.R
\name{wf_list}
\alias{wf_list}
\title{Create a list of workflows}
\usage{
wf_list(...)
}
\arguments{
\item{...}{Any number of \code{workflow} objects with names if desired.}
}
\value{
A list of class \code{wflist}.
}
\description{
Creates a list of class \code{wflist} for use with the \code{fit()} and \code{predict()} functions
in this package.
}
