% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit.wflist}
\alias{fit.wflist}
\title{Fit a list of models}
\usage{
\method{fit}{wflist}(object, data, ...)
}
\arguments{
\item{object}{A named list of \code{workflow} objects created by \code{wf_list()}.}

\item{data}{A \code{data.frame} containing data on which to fit the workflows.}

\item{...}{Arguments passed on to \code{fit.workflow()}.}
}
\value{
A named list of fitted \code{workflow} objects.
}
\description{
Fits a list of workflows to a provided \code{data.frame}.
}
