% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_multidownsample.R
\name{step_multidownsample}
\alias{step_multidownsample}
\title{Generate replicate recipes that down-sample the same data}
\usage{
step_multidownsample(.recipe, ..., .replicates)
}
\arguments{
\item{.recipe}{A \code{recipe} object.}

\item{...}{One or more selector functions to choose which variable is used
to sample the data. See \code{recipes::selections()} for more details. The
selection should result in single factor variable.}

\item{.replicates}{A positive integer indicating how many times to replicate
the recipe.}
}
\value{
A list of replicates of the provided recipe, each with a new
\code{themis::step_downsample()} step added to the sequence of existing steps (if any).
}
\description{
\code{step_multidownsample()} replicates a given recipe and adds a
\code{themis::step_downsample()} step to each replicate.
}
